Get redis-sentinel.conf file creation in place
This is just the basics so far, for sake of getting something in
place to build from.

Some assumptions are made:

* It's best to follow the model of putting constants in the
  params.pp and inheriting from that.
* A single params.pp is used (rather than creating one specifically
  for sentinel) because some existing redis params are used in sentinel.pp.

Things still to do:

* ensure service is installed and running
* ensure service is restarted on config file change
* flesh out tests
* deal with the fact that sentinel use the config file to keep state

On the latter, the most likely right thing to do is create the
config file on a different name and only copy it if there are
changes. See example from a different patchset trying the same
thing:

https://github.com/victorgp/puppet-redis/blob/f7be4187bbb77f991a53b004f8e3f2512c0c1b48/manifests/sentinel.pp#L96-L107