Make sentinel maninfest ensure the redis package
The sentinel service needs to make sure that the redis package has
been installed on the machine where sentinel is being run. All my
previous testing had assumed that there would be a redis server and
sentinel on the same machines. This doesn't have to be the case, so
the manifest needs to make allowances for that.