Enable to not manage the Redis service
In some system setup, we do not want Puppet to handle the service but
another system (ie. Pacemaker), for this reason, the service should not
be part of the catalog at all, hence the creation of this boolean.

It has to be used in conjunction of `$notify_file = false` to not depend
of the Service resource.