Fix support for EL 7 systems
The module originally tried to support EL 7 installations by utilizing
the EPEL repos with a conditional in the `preinstall` class. This
conditional was seemingly valid but failed to work correctly. This is due
to the fact that EL 7 systems have started including a version number
that has a third octet to represent build number. The conditional in
question used the `operatingsystemrelease` fact for comparison and since
the string returned could not be cast to a number, Puppet's ability to
compare would fail. In addition, the comparison of a string to a number
will not work the same way in Puppet 4 as the typing of data is more
strict. So this code will not move to the Puppet 4 DSL successfully.

This commit changes two aspects of the `preinstall` class. First it uses
the `operatingsystemmajrelease` fact that will return a single number
representing the major release of the OS. Second, since the data
returned by the fact is a string, the compared value is a string as the
greater-than and less-than comparisons in Puppet 4 can compare strings
to strings.

Note: the `operatingsystemmajrelease` fact only became available in
Facter 1.7.x and was not included in Puppet Enterprise until the 3.0
release. This does mean that this module will loose a certain amount of
backwards compatibility. Though the versions impacted are old enough
that they should not largely in use any more.