Puppet 4 is not longer an acceptable failure in CI
With the recent changes for considering EL 7 nodes, this module will be
considered to be acceptable to run in Puppet 4. In this case, we want
spec failures with Puppet 4 testing to result in failures in the TravisCI
system.

This change removes Puppet 4 as an "acceptable failure" in the TravisCI
system.