Add Sentinel IP Binding param
Redis 3.2 introduced 'protected-mode' to enhance security. The feature introduces an error in the sentinels when trying to connect from non loopback interface :

```
DENIED Redis is running in protected mode because protected mode is enabled, no bind address was specified, no authentication password is requested to clients....
```

Fix adds 'bind' to the sentinel template in order to allow the sentinels to communicate.

class { 'redis::sentinel':
  sentinel_bind => '0.0.0.0',
}
```