Add systemd service file template
If the unix socket is configured to be present under a dir in a tmpfs,
like /var/run/redis the directory must be created before launching
redis. The best way to do it is inside service file, ensure the
directory is present (and eventually security context restored) each
time redis is started.