// , Add currently parameterized values into 3.2's redis.conf
This simply takes the redis.conf file from Redis 3.2, and
adds our parameterized variables into it.

I used https://raw.githubusercontent.com/antirez/redis/3.2/redis.conf
as a starting point, and copied and replaced as needed.

This also makes sure comments don't break Travis CI's checks for
backwards incompatible items like protected-mode.

For example, if you put cluster-enabled in the comments of this template,
you get the following error from Travis:

    with parameter cluster_enabled-false

      should not contain File[/etc/redis.conf.puppet] with content =~ /cluster-enabled/ (FAILED - 2)

Note that use of the 3.2 conf is not the default.

This commit combines the addition of a configuration file option for ReDiS 3.2
with parameters for a lot of the formerly static values for the
configuration file for ReDiS 3.2.

--Nathan Basanese