// , Add new parameterized values for Redis 3.2 template
This takes any configuration value that looks like it can be set in
Redis 3.2's configuration file, and adds a parameter for it.

Values that are commented out in Redis 3.2's default configuration file
are initially set to a value of `undef`.

In this commit, many of these values will only be used in the following
template:

 * manifests/templates/redis.conf.3.2.erb

This commit also ensures that all
variables in Puppet are alphanumeric + underscores, and follows
Puppet template syntax.