Add SCL support
On Red Hat the software collections allow for multiple versions. On RHEL
this is also supported for customers where EPEL is unsupported.

To implement this, the globals pattern is implemented giving users an
easy way to select the SCL at a global level. Acceptance tests are added
to ensure it works.

For convenience, it can ensure centos-release-scl-rh is installed on
CentOS when manage_repo is true.

The custom fact to provide the server version doesn't work and also the
bolt task won't work unless the user ensured the SCL is always active.