Drop sysinit support
CentOS 6 is EOL and we can't test RHEL 6. That means there's no
guarantee it continues to work. Also, all other distros mentioned in
metadata.json run systemd. This allows the module to be cleaned up.

For the sentinel, every supported distro provides a systemd unit file as
part of the package. On Debian, it also provides a sysvinit file.
Including this in the module is not needed. In most cases, the systemd
unit file would have already shadowed the sysvinit file.