Honor redis::sentinel::package_ensure
The package_ensure parameter of the redis_sentinel class is supposed to
define status of the redis-sentinel package but in fact it is not used.
This fixes the ignored parameter.

Also, default of the parameter is updated following the recent change
of redis::package_ensure. This is required to avoid conflicting
declaration of redis package caused by a different status; 'installed'
vs 'present'.