Skip ensuring package if sentinel requires no additional package
Currently both of the redis class and the redis::sentinel class have
the package_ensure parameter which define status of the packages to
install redis and redis sentinel. However, in some distros like CentOS,
the single redis package is enough to install both redis and sentinel,
and we should set the same value to these two package_ensure parameters
to avoid conflict between two package resources maintained by these two
classes.
This change disables the package resource in the redis::sentinel class
when the single package is used, to avoid such redundant requirements.